Set the index test server shtdown port to the standard port
git-svn-id: https://ncisvn.nci.nih.gov/svn/cagrid/branches/caGrid-1_4_SHA2/Software/core@17649 dd7b0c16-3c94-4492-92de-414b6a06f0b1

+++ integration-tests/projects/indexServiceTests/src/org/cagrid/index/tests/IndexServiceSystemTest.java
@@ -87,7 +87,7 @@ public class IndexServiceSystemTest extends Story {
         // set up a testing service container
         try {
             log.debug("Creating container for testing service");
-            testServiceContainer = ServiceContainerFactory.createContainer(ServiceContainerType.TOMCAT_CONTAINER, new ContainerPorts(new Integer(8080), new Integer(8010)));
+            testServiceContainer = ServiceContainerFactory.createContainer(ServiceContainerType.TOMCAT_CONTAINER, new ContainerPorts(new Integer(8080), new Integer(8005)));
             new UnpackContainerStep(testServiceContainer).runStep();
         } catch (Throwable ex) {
             String message = "Error creating container for testing service: " + ex.getMessage();
